+++ src/main/java/cart/service/product/dto/ProductModificationDto.java
@@ -1,12 +1,14 @@
-package cart.domain.product.dto;
+package cart.service.product.dto;
 
-public class ProductCreationDto {
+public class ProductModificationDto {
+    private final Long id;
     private final String name;
     private final Integer price;
     private final String category;
     private final String imageUrl;
 
-    public ProductCreationDto(String name, Integer price, String category, String imageUrl) {
+    public ProductModificationDto(Long id, String name, Integer price, String category, String imageUrl) {
+        this.id = id;
         this.name = name;
         this.price = price;
         this.category = category;
@@ -27,5 +29,9 @@ public class ProductCreationDto {
 
     public String getImageUrl() {
         return imageUrl;
+    }
+
+    public Long getId() {
+        return id;
     }
 }
