+++ src/main/java/cart/service/user/UserService.java
@@ -1,7 +1,10 @@
-package cart.domain.user;
+package cart.service.user;
 
-import cart.domain.user.dto.UserDto;
-import cart.domain.user.dto.UserLoginDto;
+import cart.domain.user.Email;
+import cart.domain.user.Password;
+import cart.domain.user.User;
+import cart.service.user.dto.UserDto;
+import cart.service.user.dto.UserLoginDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
