+++ src/main/java/cart/web/controller/product/ProductViewController.java
@@ -1,7 +1,7 @@
 package cart.web.controller.product;
 
-import cart.domain.product.ProductService;
-import cart.domain.product.dto.ProductDto;
+import cart.service.product.ProductService;
+import cart.service.product.dto.ProductDto;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
