+++ src/test/java/cart/dao/cart/CartRepositoryImplTest.java
@@ -6,11 +6,11 @@ import cart.dao.user.UserDao;
 import cart.dao.user.UserRepositoryImpl;
 import cart.domain.TestFixture;
 import cart.domain.cart.Cart;
-import cart.domain.cart.CartRepository;
+import cart.service.cart.CartRepository;
 import cart.domain.product.Product;
-import cart.domain.product.ProductRepository;
+import cart.service.product.ProductRepository;
 import cart.domain.user.User;
-import cart.domain.user.UserRepository;
+import cart.service.user.UserRepository;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
