+++ src/test/java/cart/service/product/ProductServiceTest.java
@@ -1,7 +1,8 @@
-package cart.domain.product;
+package cart.service.product;
 
 import cart.domain.TestFixture;
-import cart.domain.product.dto.ProductDto;
+import cart.domain.product.Product;
+import cart.service.product.dto.ProductDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
